Cleanup, move module .h content into .cc files
